cleanup and fix from review
